mini-services/medical-rag-service/requirements.txt
# Medical Diagnostic RAG Service - Requirements
# Python 3.11+

# ===== Web Framework =====
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# ===== FHIR R4 Support (P12) =====
fhir.resources>=6.4.0

# ===== OpenTelemetry (P13) =====
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0

# ===== Prometheus Metrics (P13) =====
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0

# ===== Async HTTP =====
httpx==0.26.0
aiohttp==3.9.3
tenacity==8.2.3

# ===== Vector Database =====
pinecone>=3.0.0

# ===== Embeddings & ML =====
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.2.0
numpy==1.26.4
scikit-learn==1.4.0

# ===== NCBI/PubMed =====
biopython==1.83
entrezpy==2.1.3
metapub==0.5.12

# ===== P9: RAG Pipeline =====
apscheduler==3.10.4
rank-bm25==0.2.2

# ===== XML Parsing =====
lxml==5.1.0
xmltodict==0.13.0

# ===== Utilities =====
python-dotenv==1.0.1
loguru==0.7.2
orjson==3.9.12
tiktoken==0.5.2
tqdm==4.66.2

# ===== Caching =====
redis==5.0.1
aiocache==0.12.2

# ===== Security =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===== Testing =====
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# ===== Monitoring =====
prometheus-client==0.19.0

# ===== Data Processing =====
pandas==2.2.0
pyarrow==15.0.0
